// src/models/MovieDb.ts
export interface MovieDb {
    adult: boolean;
    backdrop_path: string | null;
    genre_ids: number[];
    id: number;
    original_language: string;
    original_title: string;
    overview: string;
    popularity: number;
    poster_path: string | null;
    release_date: string;
    title: string;
    video: boolean;
    vote_average: number;
    vote_count: number;
}

export interface MovieDbResult {
    page: number;
    results: MovieDb[];
    total_pages: number;
    total_results: number;
}